requirements.txt
# Vibe Coder Dependencies

# Core
PyYAML>=6.0
requests>=2.31.0

# Optional: Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Better async support
aiohttp>=3.9.0
